LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-at-user-scroll-end-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unknown source type "wheel".
CSSScrollSnap

Tests that the window can snap at user scroll end.

Scroll the page vertically and horizontally. Keep scrolling until background has turned yellow.
Press the button "Done"


Harness Error (FAIL), message = Unhandled rejection: Unknown source type "wheel".

NOTRUN Tests that window should snap at user scroll end.
